--- ui/app/pages/send/send-header/send-header.selectors.js
@@ -1,18 +1,0 @@
-import { getSelectedToken, getSendEditingTransactionId, getSendTo } from '../send.selectors.js'
-
-export function getTitleKey (state) {
-  const isEditing = Boolean(getSendEditingTransactionId(state))
-  const isToken = Boolean(getSelectedToken(state))
-
-  if (!getSendTo(state)) {
-    return 'addRecipient'
-  }
-
-  if (isEditing) {
-    return 'edit'
-  } else if (isToken) {
-    return 'sendTokens'
-  } else {
-    return 'sendETH'
-  }
-}
